requirements.dev.txt
Django==3.2
python-dotenv==1.0.0
djangorestframework==3.14.0
requests==2.28.2